Move settings to its own file
